feat(SDK-4447) moves out pushtype enum
