+++ clevertap-core/src/main/java/com/clevertap/android/sdk/pushnotification/CTPushProviderListener.java
@@ -2,6 +2,6 @@ package com.clevertap.android.sdk.pushnotification;
 
 public interface CTPushProviderListener {
 
-    void onNewToken(String token, PushConstants.PushType pushType);
+    void onNewToken(String token, PushType pushType);
 
 }
