+++ clevertap-core/src/main/java/com/clevertap/android/sdk/pushnotification/PushConstants.java
@@ -4,52 +4,6 @@ import androidx.annotation.NonNull;
 
 public interface PushConstants {
 
-    enum PushType {
-        FCM(FCM_DELIVERY_TYPE, FCM_PROPERTY_REG_ID, CT_FIREBASE_PROVIDER_CLASS, FIREBASE_SDK_CLASS);
-
-        private final String ctProviderClassName;
-
-        private final String messagingSDKClassName;
-
-        private final String tokenPrefKey;
-
-        private final String type;
-
-        PushType(
-                String type,
-                String prefKey,
-                String className,
-                String messagingSDKClassName
-        ) {
-            this.type = type;
-            this.tokenPrefKey = prefKey;
-            this.ctProviderClassName = className;
-            this.messagingSDKClassName = messagingSDKClassName;
-        }
-
-        public String getCtProviderClassName() {
-            return ctProviderClassName;
-        }
-
-        public String getMessagingSDKClassName() {
-            return messagingSDKClassName;
-        }
-
-        public String getTokenPrefKey() {
-            return tokenPrefKey;
-        }
-
-        public String getType() {
-            return type;
-        }
-
-        @NonNull
-        @Override
-        public String toString() {
-            return " [PushType:" + name() + "] ";
-        }
-    }
-
     String LOG_TAG = "PushProvider";
 
     String FCM_LOG_TAG = PushType.FCM.toString();
