+++ clevertap-core/src/main/java/com/clevertap/android/sdk/response/PushAmpResponse.java
@@ -9,7 +9,7 @@ import com.clevertap.android.sdk.CleverTapInstanceConfig;
 import com.clevertap.android.sdk.ControllerManager;
 import com.clevertap.android.sdk.Logger;
 import com.clevertap.android.sdk.db.BaseDatabaseManager;
-import com.clevertap.android.sdk.pushnotification.PushConstants.PushType;
+import com.clevertap.android.sdk.pushnotification.PushType;
 import com.clevertap.android.sdk.pushnotification.PushNotificationHandler;
 import java.util.Iterator;
 import org.json.JSONArray;
