+++ clevertap-core/src/main/java/com/clevertap/android/sdk/validation/ManifestValidator.java
@@ -17,7 +17,7 @@ import com.clevertap.android.sdk.Utils;
 import com.clevertap.android.sdk.inbox.CTInboxActivity;
 import com.clevertap.android.sdk.pushnotification.CTNotificationIntentService;
 import com.clevertap.android.sdk.pushnotification.CTPushNotificationReceiver;
-import com.clevertap.android.sdk.pushnotification.PushConstants.PushType;
+import com.clevertap.android.sdk.pushnotification.PushType;
 import com.clevertap.android.sdk.pushnotification.PushProviders;
 import java.util.ArrayList;
