+++ clevertap-hms/src/main/java/com/clevertap/android/hms/CTHmsMessageHandler.java
@@ -2,7 +2,7 @@ package com.clevertap.android.hms;
 
 import static com.clevertap.android.hms.HmsConstants.HMS_LOG_TAG;
 import static com.clevertap.android.sdk.pushnotification.PushConstants.LOG_TAG;
-import static com.clevertap.android.sdk.pushnotification.PushConstants.PushType.HPS;
+import static com.clevertap.android.sdk.pushnotification.PushType.HPS;
 
 import android.content.Context;
 import android.os.Bundle;
@@ -13,7 +13,7 @@ import com.clevertap.android.sdk.CleverTapAPI;
 import com.clevertap.android.sdk.Logger;
 import com.clevertap.android.sdk.interfaces.INotificationParser;
 import com.clevertap.android.sdk.interfaces.IPushAmpHandler;
-import com.clevertap.android.sdk.pushnotification.PushConstants.PushType;
+import com.clevertap.android.sdk.pushnotification.PushType;
 import com.clevertap.android.sdk.pushnotification.PushNotificationHandler;
 import com.clevertap.android.sdk.pushnotification.fcm.IFcmMessageHandler;
 import com.huawei.hms.push.RemoteMessage;
