+++ clevertap-hms/src/main/java/com/clevertap/android/hms/HmsPushProvider.java
@@ -1,7 +1,7 @@
 package com.clevertap.android.hms;
 
 import static com.clevertap.android.hms.HmsConstants.MIN_CT_ANDROID_SDK_VERSION;
-import static com.clevertap.android.sdk.pushnotification.PushConstants.PushType.HPS;
+import static com.clevertap.android.sdk.pushnotification.PushType.HPS;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
@@ -10,7 +10,7 @@ import androidx.annotation.RestrictTo;
 import com.clevertap.android.sdk.CleverTapInstanceConfig;
 import com.clevertap.android.sdk.pushnotification.CTPushProvider;
 import com.clevertap.android.sdk.pushnotification.CTPushProviderListener;
-import com.clevertap.android.sdk.pushnotification.PushConstants;
+import com.clevertap.android.sdk.pushnotification.PushType;
 
 /**
  * Clevertap's Huawei Plugin Ref: {@link CTPushProvider}
@@ -33,7 +33,7 @@ public class HmsPushProvider implements CTPushProvider {
 
     @NonNull
     @Override
-    public PushConstants.PushType getPushType() {
+    public PushType getPushType() {
         return HPS;
     }
